requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
motor>=3.4.0
pymongo==4.17.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.9
pydantic>=2.9.0
python-dotenv>=1.0.1
openpyxl>=3.1.2
certifi>=2024.2.2
bcrypt==3.2.2
dnspython>=2.0.0
